fix(mapper): avoid division by zero and use floating-point division in TaskList progress

+++ backend/src/main/java/se/matildaerenius/fluxo/mappers/impl/TaskListMapperImpl.java
@@ -61,6 +61,6 @@ public class TaskListMapperImpl implements TaskListMapper {
                         TaskStatus.CLOSED == task.getStatus())
                 .count();
 
-        return (double) (closedTaskCount / tasks.size());
+        return (double) closedTaskCount / (double) tasks.size();
     }
 }
